1.ota代码开始测试第一版 2.简易异常处理机制

+++ Task/Inc/bsp_iap.h
@@ -34,7 +34,7 @@ extern FileData mcuFileData;
 extern uint8_t MD5bin[512];
 /* 函数声明 ------------------------------------------------------------------*/
 extern char Judge_MD5(unsigned char* start,unsigned int len,char* output);
-extern void IAP_Write_App_Bin( uint32_t appxaddr, uint8_t * appbuf, uint32_t applen);	//在指定地址开始,写入bin
+extern void IAP_Write_App_Bin( uint32_t appxaddr, uint16_t * appbuf, uint32_t applen);	//在指定地址开始,写入bin
 extern void IAP_ExecuteApp( uint32_t appxaddr);			                              //执行flash里面的app程序
 extern int get_app_infomation(OtaData* otaInfo);
 #endif /* __IAP_H__ */
